cloudfunction/requirements.txt
langchain>=0.1.0
openai>=1.0.0
firebase-admin>=6.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=1.10.0
requests>=2.28.0 
langchain_community>=0.0.20
jsonschema
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.0,<4.0.0
pdf2image